docs(example/burst): tighten three Copilot finds — explicit int conversion, accurate median doc, accurate failures invariant
1. burst_engine.dart line 82 + burst_aggregation.dart line 43:
   add explicit .toInt() to clamp results. Dart 3 statically narrows
   int.clamp(int, int) to int so the original code compiles and runs
   correctly, but older analyzers and num-returning clamp overloads
   in mixed-numeric contexts can lose the inference. Inline comment
   on the engine site explains the defensive intent.

2. burst_aggregation.dart _median dartdoc: said 'lower of the two
   middle elements' but code returns truncated arithmetic mean.
   Reworded to match implementation.

3. burst_types.dart BurstResult.failures dartdoc: said
   'queries.length + failures.length == requested sample count'.
   Burst clamps to [1, 8] before issuing, so the invariant holds for
   the *issued* count, not the requested count. Reworded to name the
   clamp explicitly.

+++ example/lib/burst/burst_aggregation.dart
@@ -40,7 +40,7 @@ BurstResult aggregateBurst({
   // jitter_floor = max(0, p50_rtt - min_rtt) / 2; widens the per-source
   // uncertainty so a single fortunate sample cannot collapse the
   // interval below what the burst's RTT spread justifies.
-  final jitterFloor = ((medianRtt - minRtt).clamp(0, 1 << 31)) ~/ 2;
+  final jitterFloor = (medianRtt - minRtt).clamp(0, 1 << 31).toInt() ~/ 2;
   final uncertainty = minRtt ~/ 2 + jitterFloor;
 
   final spread = offsets.last - offsets.first;
@@ -61,8 +61,9 @@ BurstResult aggregateBurst({
 }
 
 /// Median of a non-empty, ascending-sorted list of integers.
-/// Even-length lists return the lower of the two middle elements
-/// (truncating-toward-zero average) to keep the return type integer.
+/// Even-length lists return the truncated arithmetic mean of the two
+/// middle elements (`(a + b) ~/ 2`) so the return type stays integer
+/// and the value is bounded by the actual sample distribution.
 int _median(List<int> sorted) {
   final n = sorted.length;
   if (n.isOdd) return sorted[n ~/ 2];
